check too long msg

+++ code/src/Application/UseCase/Bot/BotUseCase.php
@@ -149,11 +149,13 @@ final class BotUseCase
                 $this->dataEditMessageDto->message_id = $messageId;
 
                 // Удаляем все сообщения в чате по юзеру
-                $allTelegramMessageByUser = $this->telegramMessageRepository->getAllMessageByChatId($telegramUser->telegram_chat_id);
-                foreach ($allTelegramMessageByUser as $msg){
-                    $this->telegram::deleteMessage($msg->chat_id, $msg->message_id);
-                    $this->telegramMessageRepository->deleteByMessageId($msg->message_id);
-                } //
+                if ($inlineKeyboardData === 'settings_menu') {
+                    $allTelegramMessageByUser = $this->telegramMessageRepository->getAllMessageByChatId($telegramUser->telegram_chat_id);
+                    foreach ($allTelegramMessageByUser as $msg) {
+                        $this->telegram::deleteMessage($msg->chat_id, $msg->message_id);
+                        $this->telegramMessageRepository->deleteByMessageId($msg->message_id);
+                    }
+                }//
 
                 $this->telegram::editMessageTextSend($this->dataEditMessageDto);
                 $this->queueMessageRepository->deleteAllMessageByUser($telegramUser->id);
